requirements.txt
falcon
gunicorn
pyjwt
cryptography
redis
mysql-connector